backend/requirements.txt
# Video-to-Robot Processing System Requirements
# Production-level dependencies

# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Robot Control
pydobot>=1.3.2
pydobot2>=0.1.0

# Computer Vision & Hand Tracking
# IMPORTANT: Install opencv-python-headless ONLY (no opencv-python or opencv-contrib-python)
# NumPy must remain at 1.26.x because MediaPipe was compiled with NumPy 1.x
numpy==1.26.4
opencv-python-headless>=4.8.0,<4.9.0  # Headless version for server deployment
mediapipe>=0.10.0,<0.11.0

# AI & Analysis
google-genai>=1.38.0
python-dotenv>=1.0.0

# Data & Validation
pydantic>=2.0.0

# Authentication & OAuth
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0

# Database
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Development Dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
